jed pre0.99.20 changes 21-22 (small rgrep and getmail changes)
21. src/rgrep.c: Removed a misplaced return statement.
22. src/getmail.c: Removed an int typecast form an ioctl call.

+++ src/getmail.c
@@ -163,7 +163,7 @@ static void our_flock (int fd)
    f.l_start = f.l_len = 0;	       /* entire file */
 
    /* we are going to block */
-   if (fcntl (fd, F_SETLKW, (int) &f) == -1)
+   if (fcntl (fd, F_SETLKW, &f) == -1)
      {
 	exit_error ("Unable to lock file.", 0);
      }
